Rename: js-messages.ts -> messages.ts
