+++ frontend/src/components/widgets/floating-menus/ColorPicker.vue
@@ -118,7 +118,7 @@
 <script lang="ts">
 import { defineComponent, PropType } from "vue";
 
-import { RGBA } from "@/interop/js-messages";
+import { RGBA } from "@/interop/messages";
 import { hsvaToRgba, rgbaToHsva } from "@/utilities/color";
 import { clamp } from "@/utilities/math";
