+++ frontend/src/interop/messages.ts
@@ -12,7 +12,7 @@ export class JsMessage {
 }
 
 // ============================================================================
-// Add additional classes to replicate Rust's FrontendMessages and data structures below.
+// Add additional classes to replicate Rust's `FrontendMessage`s and data structures below.
 //
 // Remember to add each message to the `messageConstructors` export at the bottom of the file.
 //
