+++ frontend/src/utilities/color.ts
@@ -1,4 +1,4 @@
-import { HSVA, RGBA } from "@/interop/js-messages";
+import { HSVA, RGBA } from "@/interop/messages";
 
 export function hsvaToRgba(hsva: HSVA): RGBA {
 	const { h, s, v, a } = hsva;
